move ui state to common components module
